Fix - design review

+++ components/contacts/ContactModalRow.tsx
@@ -5,11 +5,11 @@ import { clearType, getType } from 'proton-shared/lib/contacts/property';
 import { ContactProperty, ContactPropertyChange } from 'proton-shared/lib/interfaces/contacts';
 import { classnames } from '../../helpers';
 
-import { useModals } from '../../hooks';
+// import { useModals } from '../../hooks';
 
 import ContactFieldProperty from './ContactFieldProperty';
 import ContactModalLabel from './ContactModalLabel';
-import ContactImageModal from '../../containers/contacts/modals/ContactImageModal';
+// import ContactImageModal from '../../containers/contacts/modals/ContactImageModal';
 import Icon from '../icon/Icon';
 import { OrderableHandle } from '../orderable';
 import DropdownActions from '../dropdown/DropdownActions';
@@ -40,39 +40,37 @@ const ContactModalRow = (
     }: Props,
     ref: Ref<HTMLInputElement>
 ) => {
-    const { createModal } = useModals();
-    const { field, uid } = property;
+    // const { createModal } = useModals();
+    const { field, value } = property;
     const type = clearType(getType(property.type));
-    const isImage = ['photo', 'logo'].includes(field);
-    const canEdit = false; // isImage && !!value;
+    // const isImage = ['photo', 'logo'].includes(field);
+    const canDelete = !(field === 'photo' && !value); // isImage && !!value;
 
-    const handleChangeImage = () => {
-        const handleSubmit = (value: string) => onChange({ uid, value });
-        createModal(<ContactImageModal url={property.value as string} onSubmit={handleSubmit} />);
-    };
+    // const handleChangeImage = () => {
+    //     const handleSubmit = (value: string) => onChange({ uid, value });
+    //     createModal(<ContactImageModal url={property.value as string} onSubmit={handleSubmit} />);
+    // };
 
     const list = [];
 
-    if (canEdit) {
+    // if (canEdit) {
+    //     list.push({
+    //         text: isImage ? c('Action').t`Change` : c('Action').t`Edit`,
+    //         onClick: handleChangeImage,
+    //     });
+    // }
+
+    // Delete is always available (except when primary and no image). Primary name has action row disabled.
+    if (canDelete) {
         list.push({
-            text: isImage ? c('Action').t`Change` : c('Action').t`Edit`,
-            onClick: handleChangeImage,
+            text: <Icon name="trash" className="mauto" alt={c('Action').t`Delete`} />,
+            onClick: () => {
+                if (property.uid) {
+                    onRemove(property.uid);
+                }
+            },
         });
     }
-
-    // Delete is always available. Primary name has action row disabled.
-    list.push({
-        text: canEdit ? (
-            <span className="color-global-warning">{c('Action').t`Delete`}</span>
-        ) : (
-            <Icon name="trash" className="mauto" alt={c('Action').t`Delete`} />
-        ),
-        onClick: () => {
-            if (property.uid) {
-                onRemove(property.uid);
-            }
-        },
-    });
 
     return (
         <div className="flex flex-nowrap flex-item-noshrink">
@@ -83,11 +81,9 @@ const ContactModalRow = (
                     </div>
                 </OrderableHandle>
             ) : (
-                mainItem === false && (
-                    <div className="mr0-5 flex flex-align-items-center flex-item-noshrink">
-                        <Icon name="text-justify" className="visibility-hidden" />
-                    </div>
-                )
+                <div className="mr0-5 flex flex-align-items-center flex-item-noshrink">
+                    <Icon name="text-justify" className="visibility-hidden" />
+                </div>
             )}
             <div className="flex flex-nowrap on-mobile-flex-column w100 flex-align-items-start">
                 <span
@@ -125,6 +121,7 @@ const ContactModalRow = (
                                     className={classnames([
                                         'flex flex-item-noshrink',
                                         field,
+                                        property.value,
                                         (field === 'photo' ||
                                             field === 'note' ||
                                             field === 'logo' ||
